+++ src/pages/api/admin/subcategories/delete.ts
@@ -1,14 +1,19 @@
 import { NextApiRequest, NextApiResponse } from 'next';
+import { NextApiRequestWithUser, withUser } from '../../../../libs/withUser';
 import { categoryModel } from '../../../../models/category';
 
 /**
  * To get all subcategories
  */
 async function deleteSubcategoryHandler(
-  req: NextApiRequest,
+  req: NextApiRequestWithUser,
   res: NextApiResponse
 ) {
   try {
+    const user = req.user;
+    if (!user?.id || !user.admin?.isAdmin || !user.admin?.fullAccess)
+      throw new Error('Unauthorized');
+
     const { subcategoryid } = req.query;
 
     const deletedCategories = await categoryModel.deleteSubcategory(
@@ -25,4 +30,4 @@ async function deleteSubcategoryHandler(
   }
 }
 
-export default deleteSubcategoryHandler;
+export default withUser(deleteSubcategoryHandler);
